pyproject.toml
[project]
name = "biblib"
version = "0.0.1"
description = "ADS Library Service"
authors = [{ name = "ADS Team", email = "adshelp@cfa.harvard.edu" }]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "adsmutils @ git+https://github.com/adsabs/ADSMicroserviceUtils.git@v1.3.0",
    "alembic==1.12.0",
    "psycopg2-binary==2.9.9",
    "sqlalchemy-continuum==1.3.6",
    "Flask-Mail==0.9.1",
    "Flask-Email==1.4.4",
    "Jinja2==3.1.2",
    "markupsafe==2.1.3",
    "itsdangerous==2.1.2",
    "werkzeug==2.3.8",
    "SQLAlchemy==1.3.24"
]

[project.optional-dependencies]
dev = [
    "setuptools==57.5.0",
    "Flask-Testing==0.8.1",
    "httpretty==1.1.4",
    "testing.postgresql==1.3.0",
    "pytest==6.2.5",
    "pytest-cov==3.0.0",
    "Faker==22.0.0",
    "factory-boy==3.3.0",
    "freezegun==1.4.0",
    "httmock==1.4.0",
    "mock==4.0.3",
    "flake8==4.0.1",
    "black==22.3.0",
    "isort==5.12.0",
    "coveralls==3.3.1"
]

[build-system]
requires = [
    "setuptools",
    "wheel",
    "flit_core >=3.2,<4",
    "ppsetuptools==2.0.2"
]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
addopts = "--cov=biblib --cov-report=term-missing"
testpaths = ["biblib/tests"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
